Update all type imports to use `import type` syntax Move the module augmentation for multi value to index file to avoid conflict
