+++ src/select.tsx
@@ -1,10 +1,7 @@
-import React, {
-  MutableRefObject,
-  ReactElement,
-  RefAttributes,
-  forwardRef,
-} from "react";
-import ReactSelect, { GroupBase, Props, SelectInstance } from "react-select";
+import React, { forwardRef } from "react";
+import type { MutableRefObject, ReactElement, RefAttributes } from "react";
+import ReactSelect from "react-select";
+import type { GroupBase, Props, SelectInstance } from "react-select";
 import useChakraSelectProps from "./use-chakra-select-props";
 
 type SelectType = <
